Substantially bailed on an initial design goal of allowing adjoining applicaiton integrations to authorize each request and call AppManager::run() only on authorized ones. The fundamental issue with this plan is that it makes the adjoining application a dependnecy for _all_ incoming requests, even during an update.
Instead, made it possible to split out app integrations to two components, one
running within the adjoining core that can authorize new users and one running
the .phar directly that requires a user who was already authorized.

Started documentation about writing an app integration.

The Drupal 7 upgrade test is failing because of the new use of session faux
persistence during early kernel booting; suspect this may be more of a defect
in the Drupal 7 test than in the actual src/ code.