Fix bug where session id from adjoining application was reused by Curator.
This was a security issue and also occasionally caused fatal errors on startup
of Curator, because Curator uses the native session save handlers and they have
a different set of allowed characters in session ids than e.g. the Drupal session
handler does.

The workaround involves supplying the interpreter with an alternative, secure
session ID ourselves before starting the Curator session. This is performed in
Util/SessionPrepService and tested in tests/Integration/Util/SessionPrepServiceTest.