Authenticate users in embedded mode explicitly with an IntegrationConfig flag.
This was done because the prior assumption that AppManager::applyIntegrationConfig()
would only be called if an adjoining application desired to authenticate the user
was faulty -- it is necessarily called during application startup with the default null
configuration when another integration configuration is not supplied.

Add functional test coverage for authentication status of anonymous connections.